diagnostics: Use uint64 to store memory metrics
Updated to use uint64 since memory metrics in `meminfo.c` are
stored in `unisigned long`, which is 64-bits in 64-bits linux.

BUG=b:355152864
TEST=FEATURES=test emerge-volteer diagnostics

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5746996
Tested-by: Byron Lee <byronlee@chromium.org>
Commit-Queue: Byron Lee <byronlee@chromium.org>

+++ diagnostics/cros_healthd/system/fake_meminfo_reader.cc
@@ -28,15 +28,15 @@ void FakeMeminfoReader::SetError(bool value) {
   is_error_ = value;
 }
 
-void FakeMeminfoReader::SetTotalMemoryKib(uint32_t value) {
+void FakeMeminfoReader::SetTotalMemoryKib(uint64_t value) {
   fake_total_memory_kib_ = value;
 }
 
-void FakeMeminfoReader::SetFreeMemoryKib(uint32_t value) {
+void FakeMeminfoReader::SetFreeMemoryKib(uint64_t value) {
   fake_free_memory_kib_ = value;
 }
 
-void FakeMeminfoReader::SetAvailableMemoryKib(uint32_t value) {
+void FakeMeminfoReader::SetAvailableMemoryKib(uint64_t value) {
   fake_available_memory_kib_ = value;
 }
